Update scripts to use `chrome.runtime` API, and fix connection issues
This updates react-sight to use `chrome.runtime` instead of
`chrome.extension`. I'm not sure when the API changed, or why it broke,
but this fixes the main connection issue. There are also a few weird,
explicit returns here. This is to keep the connection between the
browser tab and the background scripts alive.

Additionally, this uses the manifest to inject the content-script.