Release version 1.1.0 🚀
React-Sight is back! This should make the extenstion *actually*
useable again! It's not perfect, and there are still bugs, but it
will actually do something. By updating the webpack configuration,
dependencies, and imports, we were able to shave the file size down
by 12% (40kb)!

Lots of big changes here:
  - Fix communication issues between background script, content script,
    and browser tab.
  - Fix communication issues with react-devtools extension.
  - Revamp build system. Why did we even set it up to transpile to es5
    in the first place?
  - React sight is now
  - Upgrade all dependencies.

This one is for everyone that commented "not working" in the following
threads:
  https://github.com/React-Sight/React-Sight/issues/134
  https://github.com/React-Sight/React-Sight/issues/145