Fix circular dependency and null store errors
This fixes two major bugs:

1. Cannot read undefined from the `store` - when store was null /
undefined, the parser would crash. This initializes it to a safe value,
and wraps the parser in a try / catch.

2. Circular references in the DOM data would crash the JSON.parse call,
this fixes that by removing circular references using the `circular`
package.

Shout out to everyone who has commented "this doesn't work" since 2017!
I gotchu fam!